[IterableAppendOnlySet] add size function (#2)
We need to know the size of the set when trying to convert it into a list (e.g. to pass it to the caller of our smart contract), since memory arrays need to have a known size (variable sized arrays only work in storage).

This PR adds a view function for the size, which doesn't require any additional cost on insertion. We could alternatively store the size on the data but this would cost an sstore and every addition. Since this method is designed to be called in view functions only, we don't care about potentially running the for loop twice (once to get the count and then again to get the elements)


### Test Plan

Unit tests