Remove unbounded loop in size() implementation (#21)
More context: https://github.com/gnosis/dex-contracts/issues/641

`size()` is currently iterating over all items making it potentially very gas costly. Since this behaviour is internal to the library it is easy for dependent contracts to oversee the consequences.

This PR therefore changes the implementation so that every insert will increment the size variable.

Due to the fact that the size variable is aligned (last takes 160 bits leaving 96 in the word), the amount of gas used on insertion barely changes (by measuring test suite with eth-gas-reporter):

||min|max|avg
|---|---|---|---
No storage| 25713|66479|56044
U96|25469|66552|53192
U256|25469|86700|64361

I also removed the `atIndex` method as it is subject to the same flaw and therefore of questionable use anyway.